+++ include/StepperControl.h
@@ -1,17 +1,14 @@
-#ifndef STEPPERFUNCTIONS_H
-#define STEPPERFUNCTIONS_H
+#ifndef STEPPERCONTROL_H
+#define STEPPERCONTROL_H
 
 #include <Arduino.h>
 #include "VariableDeclarations.h"
 
 void autoStack();
-void calibrateJoyStick();
 void changeDirection();
 void dryRun();
 void homeRail();
-void joystickControl();
-void readJoystick();
-int speedControl(long loopNr);
+void overshootPosition(int position, int numberOfSteps, int direction);
 void stallDetection();
 bool stepMotor(int stepDirection, unsigned long stepperDelay);
 void toggleStepper(bool enable);
